Remove the now unnecessary mc_test_common.h
